Add new info and add clicking functionality
New info for Wind River co-op and new time ranges for the different work
experience sections. Also, now the big logos act as buttons so you can
click on them to lock in the focused view for that technology.

There seems to be an issue when going from hovering over the Java button
to the Node.js button and vice-versa.